updated and cleaned
